phase3_application/backend/requirements.txt
# Vital Lens AI - Phase 3 Backend
# FastAPI + PostgreSQL + Authentication

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic==2.5.0
pydantic-settings==2.1.0

# OAuth & 2FA
email-validator==2.1.0
pyotp==2.9.0
qrcode==7.4.2

# Data Validation & Serialization
marshmallow==3.20.1
python-dateutil==2.8.2

# Environment & Config
python-dotenv==1.0.0

# Caching & Sessions
redis==5.0.1

# Task Queue
celery==5.3.4
kombu==5.3.5

# Encryption
cryptography==41.0.7
PyCryptodome==3.19.0

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# Linting & Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Rate Limiting
slowapi==0.1.9

# Time Zone Handling
pytz==2023.3

# PDF Generation
reportlab==4.0.7
pypdf==3.17.1

# Image Processing
Pillow==10.1.0
# opencv-python==4.8.1.78  # Uncomment when implementing image preprocessing

# ML/AI Integration
numpy==1.26.2
# tensorflow==2.14.0  # Uncomment when implementing model inference
# torch==2.1.1       # Uncomment when implementing model inference

# Async HTTP
aiofiles==23.2.1

# Utilities
click==8.1.7
